functions/requirements.txt
# Do not include azure-functions-worker in this file
# The Python Worker is managed by the Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions
azure-cosmos>=4.6.0
langgraph
langsmith
langchain-openai
python-dotenv
requests
openai
